+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in 10.6.2
 - Fix hide File input if cannot edit in Accounting/functions.inc.php & Student_Billing/functions.inc.php
 - SQL use CONCAT() instead of pipes || for MySQL compatibility in Accounting/functions.inc.php & Student_Billing/functions.inc.php
 - Fix SQL use JOIN instead of WHERE EXISTS for Discipline Widgets in Widget.php
+- MySQL fix PHP fatal error when updating Course's Marking Period in Courses.php
 
 Changes in 10.6.1
 -----------------